AIModule/requirements.txt
torch==2.1.1
openai-whisper==20231117
transformers==4.35.2
googletrans==3.1.0a0
spacy==3.7.2
pytextrank==3.2.5
moviepy==1.0.3
google-cloud-storage==2.14.0
gTTS==2.4.0
regex==2023.10.3
Flask==3.0.0
